requirements.txt
streamlit
pandas
numpy
plotly
arch
gspread
openpyxl
oauth2client
python-dotenv